kactl: add support for 'failover' and 'recover' to control VRRP routers
In some cases we might need to force a master node to become a backup
node or vice versa, to make the backup node master.

The simpliest way to do it is to force master's node priority to 0.

This way the backup node will become a master due to the reception of
a priority of 0 that triggers an artificial timeout in backup state.
The master node then becomes a backup node since it receives a bigger
priority once the previous backup has become master.

The action that forces nodes' priority to 0 is called "failover", while
the action that will set the node's real priority is called "recover".

As of now, this feature is not compatible with VRRP sync groups since
they disable the VRRP priority update thread that is used to force
priority to 0.

Signed-off-by: Dimitar Delov <ddelov@arkoon.net>
Signed-off-by: Marc Finet <mfinet@arkoon.net>
Signed-off-by: Sylvain Bourdos <sbourdos@arkoon.net>