vrrp: allow group to use priority with 'global_tracking' group keyword
Consider the following architecture:

                       +----------+
                       |  client  |
                       +----------+
                             |
                       +----------+
                       | switch 1 |
                       +----------+
               VIP1  / eth0    eth0 \
             +--------+            +--------+
             | MASTER |            | BACKUP |
             +--------+            +--------+
               VIP2  \ eth1    eth1 /
                       +----------+
                       | switch 2 |
                       +----------+
                             |
                       +----------+
                       |  server  |
                       +----------+

The two VRRP instances that own VIP1 and VIP2 are grouped. In order to
address a failure between switch 1 and MASTER, floating priority is
used.  Both track the two primary interfaces eth0 and eth1 to avoid
routing problems between client and server. However, tracking is
currently not compatible with floating priority.

This patch introduces a new keyword "global_tracking" in group statement to
activate priority/script tracking on a group to determine instance state.

To use global_tracking keyword on a group, all instances synced in
this group must share the same tracking configuration (scripts and interfaces)

Signed-off-by: Dimitar Delov <ddelov@arkoon.net>
Signed-off-by: Marc Finet <mfinet@arkoon.net>
Signed-off-by: Sylvain Bourdos <sbourdos@arkoon.net>